move codes for direct encoding
